@@ -1,11 +1,11 @@
 ﻿using IPRehabModel;
 using IPRehabRepository.Contracts;
-using PatientModel;
+using PatientModel_TreatingSpecialty;
 //using PatientModel_TreatingSpecialty;
 
 namespace IPRehabRepository
 {
-    public class AnswerRepository : RepositoryBase<tblAnswer>, IAnswerRepository
+    public class AnswerRepository : RepositoryBase_IPRehab<tblAnswer>, IAnswerRepository
     {
         public AnswerRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -13,7 +13,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class CodeSetRepository : RepositoryBase<tblCodeSet>, ICodeSetRepository
+    public class CodeSetRepository : RepositoryBase_IPRehab<tblCodeSet>, ICodeSetRepository
     {
         public CodeSetRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -21,7 +21,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class EpisodeOfCareRepository : RepositoryBase<tblEpisodeOfCare>, IEpisodeOfCareRepository
+    public class EpisodeOfCareRepository : RepositoryBase_IPRehab<tblEpisodeOfCare>, IEpisodeOfCareRepository
     {
         public EpisodeOfCareRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -29,7 +29,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class PatientRepository : RepositoryBase<tblPatient>, IPatientRepository
+    public class PatientRepository : RepositoryBase_IPRehab<tblPatient>, IPatientRepository
     {
         public PatientRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -37,7 +37,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class TreatingSpecialtyPatientRepository : RepositoryBase<vTreatingSpecialtyRecent3Yrs>, ITreatingSpecialtyPatientRepository
+    public class TreatingSpecialtyPatientRepository : RepositoryBase_IPRehab<vTreatingSpecialtyRecent3Yrs>, ITreatingSpecialtyPatientRepository
     {
         public TreatingSpecialtyPatientRepository(IPRehabContext repositoryContext)
               : base(repositoryContext)
@@ -45,7 +45,23 @@ namespace IPRehabRepository
         }
     }
 
-    public class QuestionRepository : RepositoryBase<tblQuestion>, IQuestionRepository
+    public class TreatingSpecialtyDirectPatientRepository : RepositoryBase_TreatingSpecialty<RptRehabDetails>, ITreatingSpecialtyDirectPatientRepository
+    {
+        public TreatingSpecialtyDirectPatientRepository(DMTreatingSpecialtyContext repositoryContext)
+              : base(repositoryContext)
+        {
+        }
+    }
+
+    public class TreatingSpecialtyPatientDemographicRepository : RepositoryBase_TreatingSpecialty<PatientDemographic>, ITreatingSpecialtyPatientDemographicRepository
+    {
+        public TreatingSpecialtyPatientDemographicRepository(DMTreatingSpecialtyContext repositoryContext)
+              : base(repositoryContext)
+        {
+        }
+    }
+
+    public class QuestionRepository : RepositoryBase_IPRehab<tblQuestion>, IQuestionRepository
     {
         public QuestionRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -53,7 +69,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class QuestionInstructionRepository : RepositoryBase<tblQuestionInstruction>, IQuestionInstructionRepository
+    public class QuestionInstructionRepository : RepositoryBase_IPRehab<tblQuestionInstruction>, IQuestionInstructionRepository
     {
         public QuestionInstructionRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -61,7 +77,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class SignatureRepository : RepositoryBase<tblSignature>, ISignatureRepository
+    public class SignatureRepository : RepositoryBase_IPRehab<tblSignature>, ISignatureRepository
     {
         public SignatureRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -69,7 +85,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class UserRepository : RepositoryBase<tblUser>, IUserRepository
+    public class UserRepository : RepositoryBase_IPRehab<tblUser>, IUserRepository
     {
         public UserRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
@@ -77,7 +93,7 @@ namespace IPRehabRepository
         }
     }
 
-    public class QuestionMeasureRepository : RepositoryBase<tblQuestionMeasure>, IQuestionMeasureRepository
+    public class QuestionMeasureRepository : RepositoryBase_IPRehab<tblQuestionMeasure>, IQuestionMeasureRepository
     {
         public QuestionMeasureRepository(IPRehabContext repositoryContext)
             : base(repositoryContext)
